requirements.txt
# development
pymongo
flask
python-dotenv
certifi
opencv-contrib-python
face_recognition